+++ migrations/m190307_114035_tehdoc_oth_tbl.php
@@ -18,7 +18,7 @@ class m190307_114035_tehdoc_oth_tbl extends Migration
     }
     $this->createTable(self::TABLE_NAME, [
       'id' => $this->primaryKey(),
-      'eq_id' => $this->integer()->notNull(),
+      'eq_id' => $this->bigInteger()->notNull(),
       'eq_oth_title_on' => $this->boolean()->defaultValue(0),
       'eq_oth_title' => $this->string(255),
       'valid' => $this->boolean()->defaultValue(1),
